+++ leia-processor/src/main/java/com/grookage/leia/mux/DefaultMessageProcessor.java
@@ -28,10 +28,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
 
@@ -58,7 +55,7 @@ public class DefaultMessageProcessor implements MessageProcessor {
         this.executorFactory = executorFactory;
     }
 
-    protected boolean validBackends(List<String> backends) {
+    protected boolean validBackends(Set<String> backends) {
         return null != backends && !backends.isEmpty();
     }