+++ leia-processor/src/main/java/com/grookage/leia/mux/resolver/TagBasedNameResolver.java
@@ -20,11 +20,9 @@ import com.grookage.leia.models.mux.LeiaMessage;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Locale;
-import java.util.Objects;
+import java.util.*;
 import java.util.function.Supplier;
+import java.util.stream.Collectors;
 
 @AllArgsConstructor
 @Data
@@ -33,24 +31,29 @@ public class TagBasedNameResolver implements BackendNameResolver {
     private static final String BACKEND_TAG = "backend";
     private static final String TAG_SEPARATOR = "-";
 
-    private final Supplier<List<String>> backends;
+    private final Supplier<Set<String>> backends;
 
     @Override
-    public List<String> getEligibleBackends(LeiaMessage leiaMessage) {
+    public Set<String> getEligibleBackends(LeiaMessage leiaMessage) {
         final var tags = leiaMessage.getTags();
         if (null == tags || tags.isEmpty()) {
-            return List.of();
+            return Set.of();
         }
         final var backendTag = tags.stream()
                 .filter(each -> each.contains(BACKEND_TAG)).findFirst().orElse(null);
         if (null == backendTag) {
-            return List.of();
+            return Set.of();
         }
-        final List<String> eligibleBackends = null == backends ? List.of() :
-                Objects.requireNonNullElse(backends.get(), List.of());
-        final var configuredBackends = Arrays.asList(backendTag.toUpperCase(Locale.ROOT).substring(backendTag.lastIndexOf(TAG_SEPARATOR) + 1).split("\\s*::\\s*"));
+        final Set<String> eligibleBackends = null == backends ? Set.of() :
+                Objects.requireNonNullElse(backends.get(), Set.<String>of()).stream().map(String::toUpperCase)
+                        .collect(Collectors.toSet());
+        final var configuredBackends = Arrays.stream(
+                backendTag.toUpperCase(Locale.ROOT)
+                        .substring(backendTag.lastIndexOf(TAG_SEPARATOR) + 1)
+                        .split("\\s*::\\s*"))
+                .collect(Collectors.toSet());
         return eligibleBackends.stream()
                 .filter(configuredBackends::contains)
-                .toList();
+                .collect(Collectors.toSet());
     }
 }